+++ crates/voxels/src/data/surface_nets.rs
@@ -95,7 +95,7 @@ pub fn get_surface_nets(
   colors: &Vec<[f32; 3]>,
   scale: f32,
   key: [i64; 3],
-  lod: u8,
+  lod: usize,
 ) -> MeshData {
   let voxel_start = 0;
   let voxel_end = octree.get_size();
