+++ examples/editor/editor.rs
@@ -108,7 +108,7 @@ fn main() {
             [0.0, 0.2, 0.0],
             [0.0, 0.4, 0.0],
           ],
-          vec![0, range, 4, 8, 12],
+          vec![0, range as u32, 4, 8, 12],
         ))
         .add_plugin(data::CustomPlugin)
         // .add_plugin(physics::CustomPlugin)
