+++ examples/editor/graphics/mod.rs
@@ -70,7 +70,7 @@ impl Default for GraphicsResource {
 #[derive(Component)]
 pub struct ChunkGraphics {
   pub key: [i64; 3],
-  pub lod: u8,
+  pub lod: usize,
 }
 
 #[derive(Component, Clone)]
